pkcs11: merged in changes from 4.0 branch

+++ pkcs11/src/asn1.c
@@ -1,7 +1,7 @@
 /* ****************************************************************************
 
  * eID Middleware Project.
- * Copyright (C) 2008-2010 FedICT.
+ * Copyright (C) 2008-2011 FedICT.
  *
  * This is free software; you can redistribute it and/or modify it
  * under the terms of the GNU Lesser General Public License version
@@ -215,8 +215,10 @@ for (iIndex=1; iIndex < iPath; iIndex++)
 		  return (E_ASN_ITEM_NOT_FOUND);
 		}
 	}
+//p_cDat can never be bigger then p_cEnd (p_cEnd=p_cInDat + iInLen - 1;) 
+//so p_cDat-p_cInDat < iInLen - 1
 *pp_cOutDat = p_cDat;
-*l_out = iInLen - (*pp_cOutDat-p_cInDat);
+*l_out = iInLen - (unsigned int)(*pp_cOutDat-p_cInDat);
 
 return (0);
 }
